+++ flutter/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.2.0
+
+iOS: updated SprenVision 1.2.0 dependency
+Android: Added alpha dependencies:
+
+    - com.spren.sprencore:1.0.0
+    - com.spren.sprencapture:1.0.0
+
 ## 1.1.0
 
 Added 60s reading length support